Add profil

+++ models/BoardModel.php
@@ -32,7 +32,6 @@ class BoardModel extends Model{
         $statement = $this->_connectDB->prepare('SELECT Project_name FROM PROJECT where PK_Project_id=:id');
         $statement->bindValue(':id', $id);
         $res = $statement->execute();
-        
         return $res->fetchArray();
      }
 
@@ -58,9 +57,10 @@ class BoardModel extends Model{
         $req->execute();
      }
 
-     public function removeTask(string $id_task){
-      $req = $this->_connectDB->prepare('DELETE FROM TASK WHERE PK_Task_id=:id_task');
+     public function removeTask(string $id_task, string $id_user){
+      $req = $this->_connectDB->prepare('DELETE FROM TASK WHERE PK_Task_id=:id_task AND FK_Task_project IN (Select PK_Project_id FROM PROJECT WHERE FK_Project_owner=:id_user)');
       $req->bindValue(':id_task', $id_task);
+      $req->bindValue(':id_user', $id_user);
       $req->execute();
    }
